fixup - extract common query functionality to inheritable class
