+++ app/services/stats/management_information/queries/af2_high_value_query.rb
@@ -9,33 +9,8 @@
 #
 module Stats
   module ManagementInformation
-    class Af2HighValueQuery
-      include JourneyQueryable
-      include ClaimTypeQueryable
-
-      attr_reader :scheme
-
-      def self.call(options = {})
-        new(options).call
-      end
-
-      def initialize(options = {})
-        @scheme = options[:scheme]&.to_s&.upcase
-        @day = options[:day]
-        raise ArgumentError, 'scheme must be "agfs" or "lgfs"' if @scheme.present? && %w[AGFS LGFS].exclude?(@scheme)
-      end
-
-      def call
-        prepare
-        ActiveRecord::Base.connection.execute(query)
-      end
-
+    class Af2HighValueQuery < BaseQuery
       private
-
-      def prepare
-        ActiveRecord::Base.connection.execute(drop_journeys_func)
-        ActiveRecord::Base.connection.execute(create_journeys_func)
-      end
 
       def query
         <<~SQL
