+++ app/services/stats/management_information/queries/base_query.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+module Stats
+  module ManagementInformation
+    class BaseQuery
+      include JourneyQueryable
+      include ClaimTypeQueryable
+
+      attr_reader :scheme
+
+      def self.call(options = {})
+        new(options).call
+      end
+
+      def initialize(options = {})
+        @scheme = options[:scheme]&.to_s&.upcase
+        @day = options[:day]
+        raise ArgumentError, 'scheme must be "agfs" or "lgfs"' if @scheme.present? && %w[AGFS LGFS].exclude?(@scheme)
+      end
+
+      def call
+        prepare
+        ActiveRecord::Base.connection.execute(query)
+      end
+
+      private
+
+      def prepare
+        ActiveRecord::Base.connection.execute(drop_journeys_func)
+        ActiveRecord::Base.connection.execute(create_journeys_func)
+      end
+
+      def query
+        raise RuntimeError, 'query not implemented!'
+      end
+    end
+  end
+end
